Changed decimal to float. In real case scenario we will handle 2 or 3 digits after coma.

+++ LPGSavings/Domain/Car.cs
@@ -17,19 +17,19 @@ namespace LPGSavings.Domain
         private Car() {
             _fuelingHistory = new List<FuelingEntry>();
         }
-        public Car(uint distance, uint firstDistance, uint distanceLPG, uint firstDistanceLPG, decimal systemPrice, decimal systemCapacity,DateTime installationDate) : this()
+        public Car(uint distance, uint firstDistance, uint distanceLPG, uint firstDistanceLPG, float systemPrice, float systemCapacity,DateTime installationDate) : this()
         {
             Odometer = new Odometer(distance, firstDistance, distanceLPG, firstDistanceLPG);
             LPG = new LPGSystem(systemPrice,systemCapacity, installationDate);
         }
 
-        public void AddServiceEntry(ServiceType serviceType, DateTime dateOfOccure, uint odometerValue, decimal price, string description)
+        public void AddServiceEntry(ServiceType serviceType, DateTime dateOfOccure, uint odometerValue, float price, string description)
         {
             var serviceEntry = new ServiceEntry(serviceType,dateOfOccure,odometerValue,price,description);
             _serviceHistory = new List<ServiceEntry>(_serviceHistory) { serviceEntry };
         }
 
-        public void AddFueling(decimal litersLPG, decimal priceLPG, decimal litersPB, decimal pricePB, uint odometer, DateTime dateTime)
+        public void AddFueling(float litersLPG, float priceLPG, float litersPB, float pricePB, uint odometer, DateTime dateTime)
         {
             var fuelingEntry = new FuelingEntry(litersLPG, priceLPG, litersPB, pricePB, odometer, dateTime);
             _fuelingHistory = new List<FuelingEntry>(_fuelingHistory) { fuelingEntry };
