+++ LPGSavings/Domain/LPGSystem.cs
@@ -5,12 +5,12 @@ namespace LPGSavings.Domain
     public class LPGSystem
     {
         public int LPGSystemId { get; private set; }
-        public decimal Price { get; private set; }
-        public decimal Capacity { get; private set; }
+        public float Price { get; private set; }
+        public float Capacity { get; private set; }
         public DateTime InstallationDate { get; private set; }
 
         private LPGSystem() { } 
-        public LPGSystem(decimal price, decimal capacity, DateTime installationDate)
+        public LPGSystem(float price, float capacity, DateTime installationDate)
         {
             Price = price;
             Capacity = capacity;
