+++ LPGSavings/Services/CarService.cs
@@ -10,7 +10,7 @@ namespace LPGSavings.Services
     public class CarService : ICarService
     {
 
-        public async Task AddService(ServiceType serviceType, DateTime dateOfOccure, uint odometerValue, decimal price, string description)
+        public async Task AddService(ServiceType serviceType, DateTime dateOfOccure, uint odometerValue, float price, string description)
         {
             using var context = new MainContext();
             var car = context.Cars.Include(a => a.ServiceHistory).First();
@@ -24,7 +24,7 @@ namespace LPGSavings.Services
             return await context.Cars.AsNoTracking().FirstOrDefaultAsync().ConfigureAwait(false);
         }
 
-        public async Task InitializeCar(uint distance, uint distanceLPG, decimal systemPrice, decimal systemCapacity, DateTime installationDate)
+        public async Task InitializeCar(uint distance, uint distanceLPG, float systemPrice, float systemCapacity, DateTime installationDate)
         {
             using var context = new MainContext();
             if (context.Cars.Count() > 0)
